+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - add integer only support
 - add escape of ffmpeg strings
 - add progressbar when parsing episodes
+- add direct arguments for login and download
 - try signing in when 403 is thrown
 
 # 1.0.0rc5